Added stub encoder funcs

+++ Source/PopH264.cpp
@@ -283,8 +283,8 @@ __export void PopH264_EncoderPeekData(int32_t Instance,char* MetaJsonBuffer,int3
 			auto& Encoder = PopH264::EncoderInstanceManager.GetInstance(Instance);
 
 			//	add generic meta
-			MetaJson["InputQueueCount"] = Encoder.GetFrameQueueCount();
-			MetaJson["OutputQueueCount"] = Encoder.GetPacketQueueCount();
+			MetaJson["InputQueueCount"] = static_cast<int32_t>(Encoder.GetFrameQueueCount());
+			MetaJson["OutputQueueCount"] = static_cast<int32_t>(Encoder.GetPacketQueueCount());
 			
 			Encoder.PeekPacket(MetaJson);
 		}
